model을 domain과 service단으 나누어 domain에는 사용할 객체를 생성할 클래스와 자잘한 기능들을 넣어두고 service에 게임에 필요한 각각의 비즈니스 로직을 만듬. 그리고 Controller는 service단의 로직과 View를 사용하여 게임을 진행함
