+++ src/main/java/racingcar/domain/CreateCarList.java
@@ -1,13 +1,13 @@
-package racingcar.model;
+package racingcar.domain;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class CreateCar {
+public class CreateCarList {
     private final List<Car> cars = new ArrayList<>();
 
-    public List<Car> createCar(String string_cars){
-        String[] car_name = settingCar(string_cars);
+    public List<Car> createCar_list(String string_cars){
+        String[] car_name = splitCarsName(string_cars);
 
         for (String s : car_name) {
             cars.add(new Car(s));
@@ -16,7 +16,7 @@ public class CreateCar {
         return cars;
     }
 
-    public String[] settingCar(String cars){
+    public String[] splitCarsName(String cars){
     String[] car = cars.split(",");
     ValidCheck.namesize_check(car);
